chore: format code
